pyproject.toml
[tool.poetry]
name = "c4-discord-bot"
version = "0.1.0"
description = ""
authors = ["Luna <itsactuallyluna9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.4.0"
httpx = "^0.28.1"
python-dotenv = "^1.0.1"
apscheduler = "^3.11.0"
pillow = "^11.1.0"
numpy = "1.26.4"
pandas = "^2.2.3"
joblib = "^1.3.0"
keras = "^3.8.0"
google-generativeai = "^0.8.4"
torch = "^2.6.0"
nltk = "^3.9.1"
matplotlib = "^3.10.0"
scikit-learn = {path = "scikit_learn-1.2.2-cp312-cp312-macosx_14_0_arm64.whl"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
